+++ CHANGELOG.md
@@ -9,3 +9,4 @@
 - [feature: optimize polaris-ratelimit-example, add caller query params and headers, add callee custom label resolver.](https://github.com/Tencent/spring-cloud-tencent/pull/951)
 - [refactor:refactor reactor code in router module.](https://github.com/Tencent/spring-cloud-tencent/pull/955)
 - [refactor: refactor circuitbreaker .](https://github.com/Tencent/spring-cloud-tencent/pull/963)
+- [feature: add polaris ring hash load balancer .](https://github.com/Tencent/spring-cloud-tencent/pull/970)